+++ markets/bfp-market/contracts/interfaces/IPerpMarketFactoryModule.sol
@@ -45,11 +45,11 @@ interface IPerpMarketFactoryModule is IMarket, IBasePerpMarket {
         /// Current oracle price (not accounting for pd adjustments).
         uint256 oraclePrice;
         /// Current rate of funding velocity.
-        int256 fundingVelocity;
+        int128 fundingVelocity;
         /// Current funding rate as a function of funding velocity.
-        int256 fundingRate;
+        int128 fundingRate;
         /// Current utilization rate
-        uint256 utilizationRate;
+        uint128 utilizationRate;
         /// Amount of size remaining last recorded in current window.
         uint256 remainingLiquidatableSizeCapacity;
         /// block.timestamp of when the last liqudation had occurred.
